Move cargo manifest package and workspace types
